+++ packages/dialog/stories/dialog-wrapper.stories.ts
@@ -672,3 +672,8 @@ lazyHero.args = {
 lazyHero.swc_vrt = {
     skip: true,
 };
+
+lazyHero.parameters = {
+    // Disables Chromatic's snapshotting on a global level
+    chromatic: { disableSnapshot: true },
+};